Error status (#106)
* update validation of state

* update & refactor analysis mailer

* rescue and update state for prediction

* update & refactor evaluation mailer

* add error and create helper

* add error and create prediction helper

* add error state to evaluation view

* update models spec

* update view spec

* raise if script failer

* update CHANGELOG.md

+++ app/jobs/prediction_job.rb
@@ -14,12 +14,15 @@ class PredictionJob < ActiveJob::Base
     end
 
     args = [prediction_id, prediction.model, Settings.prediction.tmp_file_name]
-    system "Rscript #{Rails.root}/scripts/predict.r #{args.join(' ')}"
+    is_success = system "Rscript #{Rails.root}/scripts/predict.r #{args.join(' ')}"
+    raise StandardError unless is_success
 
     YAML.load_file(Rails.root.join(data_dir, 'prediction.yml')).each do |number, result|
       prediction.results.create!(number: number) if result == 1
     end
     FileUtils.rm_rf(data_dir)
     prediction.update!(state: 'completed')
+  rescue StandardError
+    prediction.update!(state: 'error')
   end
 end
